test(KEN-935): the rename tests get their own file back [no-changelog]
The rename-and-declaration tests move to author/import/tests/rename.rs,
the seam the size ratchet names as its first remedy for a file over its
class. The Copilot-driven refusal for a second agent format joins them
there rather than staying in review.rs, which it only reached because
tests.rs could not grow.

That file is 524 lines now, back under the 800-line threshold of its
class, so its size-ratchet baseline row is gone rather than left standing
for a file that no longer earns one.

The four comments trimmed to fit the cap are back as they were: why
`checked` asserts an item count at all, why an escape-bearing candidate
name reaches the refusal, what a rewrite reaching a body file would cost,
and what the bare tree in the fixture is for. A size cap is the wrong
thing to be deciding which explanations survive.
