+++ tools/size-ratchet-baseline.tsv
@@ -1,6 +1,5 @@
 .github/workflows/review-gate-writer.yml	662
 .github/workflows/skill-tests.yml	551
-crates/core/src/author/import/tests.rs	815
 crates/core/src/engine/detach.rs	417
 crates/core/src/error.rs	503
 crates/core/src/mapping.rs	411
